appointment_management/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.8
django-webpush==0.3.3
http-ece==1.1.0
idna==3.3
py-vapid==1.8.2
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.3
pywebpush==1.9.4
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
